Move REST and SOTW to external servers
